+++ x/finality/keeper/grpc_query.go
@@ -97,7 +97,7 @@ func (k Keeper) Block(ctx context.Context, req *types.QueryBlockRequest) (*types
 		return nil, err
 	}
 
-	return &types.QueryBlockResponse{Block: b}, nil
+	return &types.QueryBlockResponse{Block: convertToIndexedBlockResponse(b)}, nil
 }
 
 // ListBlocks returns a list of blocks at the given finalisation status
@@ -129,7 +129,7 @@ func (k Keeper) ListBlocks(ctx context.Context, req *types.QueryListBlocksReques
 	}
 
 	resp := &types.QueryListBlocksResponse{
-		Blocks:     ibs,
+		Blocks:     convertToIndexedBlockResponseList(ibs),
 		Pagination: pageRes,
 	}
 	return resp, nil
@@ -289,6 +289,22 @@ func convertToSigningInfosResponse(signInfos []types.FinalityProviderSigningInfo
 		response[i] = convertToSigningInfoResponse(info)
 	}
 	return response
+}
+
+func convertToIndexedBlockResponse(ib *types.IndexedBlock) *types.IndexedBlockResponse {
+	return &types.IndexedBlockResponse{
+		Height:    ib.Height,
+		AppHash:   ib.AppHash,
+		Finalized: ib.Finalized,
+	}
+}
+
+func convertToIndexedBlockResponseList(ibs []*types.IndexedBlock) []*types.IndexedBlockResponse {
+	var blockResponses []*types.IndexedBlockResponse
+	for _, ib := range ibs {
+		blockResponses = append(blockResponses, convertToIndexedBlockResponse(ib))
+	}
+	return blockResponses
 }
 
 func convertToEvidenceResponse(evidence *types.Evidence) *types.EvidenceResponse {