+++ CHANGELOG.md
@@ -77,6 +77,7 @@ generalized unbonding handler
 * [#202](https://github.com/babylonlabs-io/babylon/pull/202) Adjusted handling of `FinalityProviderWithMeta` in btcstaking keeper queries to improve API security.
 * [#203](https://github.com/babylonlabs-io/babylon/pull/203) Adjusted handling of `RewardGauge` in incentive keeper queries to improve API security.
 * [#208](https://github.com/babylonlabs-io/babylon/pull/208) Adjusted handling of `Evidence` in finality keeper queries to improve API security.
+* [#209](https://github.com/babylonlabs-io/babylon/pull/209) Adjusted handling of `IndexedBlock` in finality keeper queries to improve API security.
 
 ### State Machine Breaking